requirements.txt
tqdm==4.66.2
seaborn==0.12.2
scikit-learn==1.3.1
py7zr==0.20.8
PyMySQL==1.1.0
peft==0.10.0
pandas==2.0.2
nvitop
javalang==0.13.0
gensim==4.3.2
flask==2.3.2
Django==4.2.2
cryptography==41.0.7
chardet==5.2.0
celery-progress==0.3
rarfile==4.1
pdfplumber==0.10.3
pycryptodome==3.18.0
reportlab==4.0.4
openpyxl==3.1.2
xlwt==1.3.0
celery==5.3.1
numpy==1.24.3
DateTime==5.2
tokenizers==0.19.1
transformers==4.40.2
datasets
bitsandbytes
tree_sitter
vllm==0.5.3